Refactor querying scripts and combine search results.
The first change is that profile_query_script.py and table_query_script.py
were refactored to have the same structure. This was primarily to separate
the search and score algorithms from I/O (taking in the query and printing
results). This refactoring allows its functions to be called independently
in query_script.py. Moreover, it makes clearer the purpose of both scripts
(running queries from the command line) by renaming them.

Secondly, table_query_script.py had its compute_score function updated, to
match the profile_query_script.py / score.py function and return scores in
a comparable range.

Finally, query_script.py takes advantage of the refactoring and is able to
search for both tables and profiles matching a given query. Because scores
are all in the same range, it simply is able to run both scripts and build
combined results from the two independent sets of results.