Improve profile search behavior.
Eliminated removal of stopwords on profiles, so that names such as
"Columbus, IN" did not get parsed to "Columbus" (IN was treated as
the preposition "in"). This means that searches for "Columbus, IN"
will actually return that as the top result (and likewise for more
places in Indiana).

Fixed a bug with sorting of results where sumlevels were being sorted
alphabetically rather than numerically.