Update database with topics entries
The functions created before now update search_metadata with topic
page entries. Data for the topic name, relevant tables, and text on
the page is included.

The data can be re-updated simply as "python topic_scraper.py", as
each run of the script deletes the old topics entries before adding
the new ones in, refreshing the table.