Remove broken query script, update docs
Removed query_script.py, as there have been many changes to the API
and to the database that were not reflected in the script. Furthermore,
it was reliant on the individual profile and table query scripts, which
themselves have since been deleted. Because of all this, and the fact
that its original purpose was only for testing, there is no reason to
keep it around.

full-text-guide.md has been updated to reflect this.