Update search_metadata with URL
topic_scraper.py now adds the URL to search_metadata. It was necessary
to include this in the database because there is no standard format
for topic pages URLs that *always* works, so they cannot be generated
on the fly. If the URLs change, the script will need to be re-run, but
I do not anticipate that happening.

Fixed some whitespace as well.