Add topic search to API
The full text search endpoint now returns topics as well as profiles
and tables. There is work to be done with ranking of the results -
it is difficult to rank the results to give topics precedence while
not flooding the results with topics only (for queries like 'income',
which return 11 topic pages and a ton of tables). More work to be done
refactoring as well.