Refactor score computation
Refactor compute_profile_score() and compute_table_score() into
one function, compute_score(). This allows for less reuse of code
in the score computation. In addition, the function takes a row
directly, rather than individual attributes of the row (like population
or priority), so it is resilient to changes to the query format.