Use only geoid in profile URL
The profile URLs returned only use the geoid now, rather than the
geoid and display name earlier. The Census Reporter app will interpret
a URL with geoid only correctly, and redirect to the "right" URL. This
was done to avoid having to slugify display names and have them
potentially be wrong in the future.