Detect and parse annotations to table codes
Modify topic_scraper.py to detect annotations to table codes. This
searches the characters after each table code found for annotations
(like the dagger, double dagger, etc.) and keeps track of them.
This functionality can later be used to check the validity of such
annotations.

The core usage of the script has not changed.