Lookup annotations for other table versions
This script now reads the annotations next to each table code (thanks
to the modifications to topic_scraper.py). Using these, it checks the
validity of any tables that should exist, according to the annotation
or annotations listed.

e.g., an annotation of a double dagger means the collapsed version is
present, so we check that it actually is.