Add synonyms to search_metadata table's document column.
Previously for local testing a postgres thesaurus was used, but this
feature is not deployable. Thus, after creating the search_metadata
table in the indexing sql script, we manually check for common synonyms
such as Ft. and Fort or St. and Saint, then concatenate the synonyms (cast as a
tsvector) to the document.